MAINT: stats: begin translation of FNI to UNU.RAN
