+++ benchmarks/benchmarks/stats.py
@@ -494,32 +494,6 @@ class BenchQMCHalton(Benchmark):
         seq.random(n, workers=workers)
 
 
-class NumericalInverseHermite(Benchmark):
-
-    param_names = ['distribution']
-    params = [distcont]
-
-    def setup(self, *args):
-        self.rand = [np.random.normal(loc=i, size=1000) for i in range(3)]
-
-    def time_fni(self, distcase):
-        distname, shapes = distcase
-        slow_dists = {'ksone', 'kstwo', 'levy_stable', 'skewnorm'}
-        fail_dists = {'beta', 'gausshyper', 'geninvgauss', 'ncf', 'nct',
-                      'norminvgauss', 'genhyperbolic', 'studentized_range'}
-
-        if distname in slow_dists or distname in fail_dists:
-            raise NotImplementedError("skipped")
-
-        dist = getattr(stats, distname)(*shapes)
-
-        with np.testing.suppress_warnings() as sup:
-            sup.filter(RuntimeWarning, "overflow encountered")
-            sup.filter(RuntimeWarning, "divide by zero")
-            sup.filter(RuntimeWarning, "invalid value encountered")
-            stats.NumericalInverseHermite(dist)
-
-
 class DistanceFunctions(Benchmark):
     param_names = ['n_size']
     params = [
