+++ scipy/stats/__init__.py
@@ -456,7 +456,7 @@ from .contingency import chi2_contingency
 from ._bootstrap import bootstrap
 from ._entropy import *
 from ._hypotests import *
-from ._rvs_sampling import rvs_ratio_uniforms, NumericalInverseHermite
+from ._rvs_sampling import rvs_ratio_uniforms
 from ._unuran import *  # noqa
 from ._page_trend_test import page_trend_test
 from ._mannwhitneyu import mannwhitneyu
