+++ scipy/stats/_unuran/__init__.py
@@ -2,5 +2,6 @@ from .unuran_wrapper import (  # noqa: F401
     TransformedDensityRejection,
     DiscreteAliasUrn,
     NumericalInversePolynomial,
+    NumericalInverseHermite,
     UNURANError
 )
